requirements.txt
djangorestframework==3.11.0
drf-nested-routers==0.91
psycopg2==2.8.5
pytest-django==3.9.0
sqlalchemy==1.3.16
